app/requirements.txt
Django==3.0.7
psycopg2==2.8.5
djangorestframework==3.11.0
uWSGI==2.0.19